Registered users have been set to uniqueness based on username and email Emails are case insensitive User can see its own page.
Missing still are user links to jobsearch and view of all of there jobs applied too
Automatic login(if registered for the first time user should not need to login in)
Login and Logout features from rootindex.
