+++ .byebug_history
@@ -1,4 +1,7 @@
 exit
+message
+type
+exit
 user_params
 params
 exit
