+++ app/helpers/sessions_helper.rb
@@ -1,0 +1,17 @@
+module SessionsHelper
+  def login
+  end
+
+  def remember
+
+  end
+
+  def forget
+
+  end
+
+  def logout
+
+  end 
+
+end
