+++ app/views/layouts/application.html.erb
@@ -8,6 +8,10 @@
 </head>
 <body>
 
+  <% flash.each do |type, message| %>
+  <%= content_tag :div, message, class: "flash_#{type}"%>
+  <% end %>
+
 <%= yield %>
 
 </body>
