+++ app/views/users/show.html.erb
@@ -1,1 +1,5 @@
 <h1>Welcome <%= @user.first_name %></h1>
+
+<h3>here are the jobs you have applied to _______</h3><br>
+
+<h3><%= link_to "Search for Jobs" %></h3>
